restructure forms as fragments for wizard + withLocaleHandlers for handle
