+++ src/components/Steps.js
@@ -30,29 +30,29 @@ const Steps = ({ intl, step, done }: Props) => (
       <Row>
         <Col sm="3">
           <Step
+            completed={step > 0 || done}
+            current={step === 0}
+            label={intl.messages['site.steps.berths']}
+          />
+        </Col>
+        <Col sm="3">
+          <Step
             completed={step > 1 || done}
             current={step === 1}
-            label={intl.messages['site.steps.berths']}
+            label={intl.messages['site.steps.boat_information']}
           />
         </Col>
         <Col sm="3">
           <Step
             completed={step > 2 || done}
             current={step === 2}
-            label={intl.messages['site.steps.boat_information']}
-          />
-        </Col>
-        <Col sm="3">
-          <Step
-            completed={step > 3 || done}
-            current={step === 3}
             label={intl.messages['site.steps.applicant']}
           />
         </Col>
         <Col sm="3">
           <Step
-            completed={step > 4 || done}
-            current={step === 4}
+            completed={step > 2 || done}
+            current={step === 2}
             label={intl.messages['site.steps.send_application']}
           />
         </Col>
