+++ src/components/containers/StepsContainer.js
@@ -1,0 +1,11 @@
+// @flow
+import { connect } from 'react-redux';
+import Steps from '../Steps';
+
+export default connect(
+  state => ({
+    step: state.forms.step,
+    done: state.forms.done
+  }),
+  {}
+)(Steps);
