--- src/components/containers/UserPageContainer.js
@@ -1,4 +1,0 @@
-// @flow
-import UserPage from '../pages/UserPage';
-
-export default UserPage;
