+++ src/components/pages/FormPage.js
@@ -5,10 +5,11 @@ import styled from 'styled-components';
 
 import Layout from './Layout';
 import Legend from '../Legend';
-import Steps from '../Steps';
-import Wizard from '../forms/container/WizardContainer';
-import RegisteredBoat, { registeredBoat } from '../forms/partials/RegisteredBoat';
-import UnRegisteredBoat, { unRegisteredBoat } from '../forms/partials/UnRegisteredBoat';
+import Steps from '../containers/StepsContainer';
+import Wizard from '../forms/containers/WizardContainer';
+import RegisteredBoat, { schema as RegisteredBoatSchema } from '../forms/fragments/RegisteredBoat';
+import PrivatePerson, { schema as PrivatePersonSchema } from '../forms/fragments/PrivatePerson';
+import Overview, { schema as OverviewSchema } from '../forms/fragments/Overview';
 import validation from '../../utils/formValidation';
 
 const Content = styled.div`
@@ -25,8 +26,9 @@ const BoatPage = () => (
     </Legend>
     <Content>
       <Wizard>
-        <UnRegisteredBoat validate={validation(unRegisteredBoat)} />
-        <RegisteredBoat validate={validation(registeredBoat)} />
+        <RegisteredBoat validate={validation(RegisteredBoatSchema)} />
+        <PrivatePerson validate={validation(PrivatePersonSchema)} />
+        <Overview validate={validation(OverviewSchema)} />
       </Wizard>
     </Content>
   </Layout>
