+++ src/components/pages/ThankyouPage.js
@@ -1,0 +1,7 @@
+// @flow
+import React from 'react';
+import Layout from './Layout';
+
+const BerthPage = () => <Layout>Thanks!</Layout>;
+
+export default BerthPage;
