+++ src/utils/formValidation.js
@@ -1,10 +1,11 @@
 import Joi from 'joi';
 
 export default schema => values => {
-  const { error } = Joi.validate(values, schema, { abortEarly: false });
+  const { error } = Joi.validate(values, schema, { abortEarly: false, allowUnknown: true });
 
   if (error && error.details) {
     const { details } = error;
+    console.log('ERROOR', details);
     return details.reduce(
       (reduction, current) => ({
         ...reduction,
@@ -13,5 +14,6 @@ export default schema => values => {
       {}
     );
   }
+  console.log('KAIKKI OK!');
   return {};
 };
